Create create_new_user
This script will take from a CSV file with the columns:
First Name | Last Name | username |
and then checks Canvas if the user exists, and if they don't, creates them an account, emails them that it's been generated, and creates them a sandbox course and makes them the teacher of that course. If you don't want to use a CSV you can modify the code and take off the Comment hashes at line 21-24.